Add support for slaw conses
A new JS class, Cons, simply joins two Javascript values together and provides
accessors for getting at them.  When converted to Slaw, the Cons's car and cdr
JS values are converted to slaw values, and combined into a slaw cons.

The same happens in reverse when converting a slaw cons to a Javascript value.